requirements.txt
asgiref==3.8.1
Django==5.1.1
django-restframework==0.0.1
environ==1.0
sqlparse==0.5.1
tzdata==2024.1
